Added more clients

+++ algorithms.h
@@ -3,9 +3,9 @@
 
 #include "fileio.h"
 
-void findSudokuBruteForce(int **sudoku, int row, int col, int side, ListSudoku *solved, int *cost);
+void findSudokuBruteForce(int **sudoku, int row, int col, int side, ListSudoku *solved, long long *cost);
 
-void findSudokuAdvanced(Sudoku s, ListSudoku *solved, int *cost);
+void findSudokuAdvanced(Sudoku s, ListSudoku *solved, long long *cost);
 
 int isValidPlacement(int **sudoku, int num, int row, int col, int side);
